Progress on online multiplayer -feature
Begun implementing online multiplayer, both ui and logic.
