+++ .flutter-plugins
@@ -1,6 +1,7 @@
 # This is a generated file; do not edit or check into version control.
 cloud_firestore=/Library/flutter/.pub-cache/hosted/pub.dartlang.org/cloud_firestore-0.13.5/
 cloud_firestore_web=/Library/flutter/.pub-cache/hosted/pub.dartlang.org/cloud_firestore_web-0.1.1+2/
+device_info=/Library/flutter/.pub-cache/hosted/pub.dartlang.org/device_info-0.4.2+2/
 firebase_analytics=/Library/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_analytics-5.0.11/
 firebase_core=/Library/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_core-0.4.4+3/
 firebase_core_web=/Library/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_core_web-0.1.1+2/
