+++ .idea/libraries/Dart_Packages.xml
@@ -79,10 +79,10 @@
             </list>
           </value>
         </entry>
-        <entry key="cupertino_icons">
+        <entry key="device_info">
           <value>
             <list>
-              <option value="/Library/flutter/.pub-cache/hosted/pub.dartlang.org/cupertino_icons-0.1.3/lib" />
+              <option value="/Library/flutter/.pub-cache/hosted/pub.dartlang.org/device_info-0.4.2+2/lib" />
             </list>
           </value>
         </entry>
@@ -191,6 +191,13 @@
             </list>
           </value>
         </entry>
+        <entry key="nested">
+          <value>
+            <list>
+              <option value="/Library/flutter/.pub-cache/hosted/pub.dartlang.org/nested-0.0.4/lib" />
+            </list>
+          </value>
+        </entry>
         <entry key="path">
           <value>
             <list>
@@ -226,10 +233,24 @@
             </list>
           </value>
         </entry>
+        <entry key="provider">
+          <value>
+            <list>
+              <option value="/Library/flutter/.pub-cache/hosted/pub.dartlang.org/provider-4.0.5/lib" />
+            </list>
+          </value>
+        </entry>
         <entry key="quiver">
           <value>
             <list>
               <option value="/Library/flutter/.pub-cache/hosted/pub.dartlang.org/quiver-2.0.5/lib" />
+            </list>
+          </value>
+        </entry>
+        <entry key="random_string">
+          <value>
+            <list>
+              <option value="/Library/flutter/.pub-cache/hosted/pub.dartlang.org/random_string-2.0.1/lib" />
             </list>
           </value>
         </entry>
@@ -352,7 +373,7 @@
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/collection-1.14.11/lib" />
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/convert-2.1.1/lib" />
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/crypto-2.1.3/lib" />
-      <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/cupertino_icons-0.1.3/lib" />
+      <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/device_info-0.4.2+2/lib" />
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/drawing_animation-0.1.3/lib" />
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/firebase-7.3.0/lib" />
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_analytics-5.0.11/lib" />
@@ -365,12 +386,15 @@
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/js-0.6.1+1/lib" />
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/matcher-0.12.6/lib" />
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/meta-1.1.8/lib" />
+      <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/nested-0.0.4/lib" />
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/path-1.6.4/lib" />
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/path_parsing-0.1.4/lib" />
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/pedantic-1.9.0/lib" />
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/petitparser-2.4.0/lib" />
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/plugin_platform_interface-1.0.2/lib" />
+      <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/provider-4.0.5/lib" />
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/quiver-2.0.5/lib" />
+      <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/random_string-2.0.1/lib" />
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/shared_preferences-0.5.6+3/lib" />
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/shared_preferences_macos-0.0.1+6/lib" />
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/shared_preferences_platform_interface-1.0.3/lib" />
