+++ .idea/libraries/Flutter_Plugins.xml
@@ -8,6 +8,7 @@
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_core-0.4.4+3" />
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/cloud_firestore_web-0.1.1+2" />
       <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_core_web-0.1.1+2" />
+      <root url="file:///Library/flutter/.pub-cache/hosted/pub.dartlang.org/device_info-0.4.2+2" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
