+++ lib/models.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/cupertino.dart';
+
+class WaitingRoom {
+  final String roomCode;
+  final int activePlayers;
+
+  WaitingRoom({@required this.roomCode, @required this.activePlayers});
+}
