+++ pubspec.yaml
@@ -14,7 +14,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.6.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -26,7 +26,9 @@ dependencies:
   drawing_animation: ^0.1.3
   tuple: ^1.0.3
   shared_preferences: ^0.5.6+3
-  cupertino_icons: ^0.1.3
+  random_string: ^2.0.1
+  provider: ^4.0.5
+  device_info: ^0.4.2+2
 
 dev_dependencies:
   flutter_test:
